edit virtual-list, split Search component
