+++ host/pages/api/botkube.js
@@ -6,6 +6,11 @@ export default async function handler(req, res) {
   const jsonDirectory = path.join(process.cwd(), 'json');
   //Read the json data file data.json
   const fileContents = await fs.readFile(jsonDirectory + '/botkube.json', 'utf8');
+  // Add key property
+  const updatedList = JSON.parse(fileContents).map((item, idx) => ({
+    ...item,
+    key: idx+1,
+  }));
   //Return the content of the data file in json format
-  res.status(200).json(fileContents);
+  res.status(200).json(updatedList);
 }
