+++ host/pages/botkube/virtual-list/page.jsx
@@ -1,5 +1,7 @@
+import { Input } from 'antd';
 import CustomTable from '#/components/VirtualList';
 import useSWR from 'swr';
+import { useState, useEffect } from 'react';
 
 const columns = [
   {
@@ -23,17 +25,52 @@ const columns = [
 ];
 
 const fetcher = (url) => fetch(url).then((res) => res.json());
+const useBotkubeInfo = () => {
+  const { data, error } = useSWR('/api/botkube', fetcher);
+
+  return {
+    data,
+    error,
+  };
+};
+
+const { Search } = Input;
 
 export default function Page() {
-  const { data, error } = useSWR('/api/botkube', fetcher);
+  const { data, error } = useBotkubeInfo();
+  const [dataList, setDataList] = useState([]);
+  const [filterInput, setFilterInput] = useState('')
+
+  useEffect(() => {
+    if (data) {
+      setDataList(data);
+    }
+  }, [data]);
+
   if (error) return <div>Failed to load</div>;
   if (!data) return <div>Loading...</div>;
 
+  const filterData = () => {
+    if(filterInput === '') return dataList
+
+    if(filterInput) {
+      return dataList.filter(({ summary }) => summary.includes(filterInput))
+    }
+  }
+
   return (
     <div>
+      <Search
+        size="large"
+        placeholder="Enter the keyword"
+        allowClear
+        enterButton="Search"
+        onSearch={setFilterInput}
+      />
+      <br />
       <CustomTable
         columns={columns}
-        dataList={JSON.parse(data)}
+        dataList={filterData()}
       >
       </CustomTable>
     </div>
